+++ tests/unit/test_completer.py
@@ -23,7 +23,7 @@ from awscli.completer import Completer
 LOG = logging.getLogger(__name__)
 
 GLOBALOPTS = ['--debug', '--endpoint-url', '--no-verify-ssl',
-              '--no-paginate', '--output', '--profile',
+              '--no-paginate', '--output', '--page-size', '--profile',
               '--region', '--version', '--color', '--query', '--no-sign-request']
 
 COMPLETIONS = [
@@ -62,7 +62,7 @@ COMPLETIONS = [
      set(['--filters', '--dry-run', '--no-dry-run', '--endpoint-url',
           '--no-verify-ssl', '--no-paginate', '--no-sign-request',
           '--output', '--profile', '--starting-token', '--max-items',
-          '--region', '--version', '--color', '--query'])),
+          '--region', '--version', '--color', '--query', '--page-size'])),
     ('aws s3', -1, set(['cp', 'mv', 'rm', 'mb', 'rb', 'ls', 'sync', 'website'])),
     ('aws s3 m', -1, set(['mv', 'mb'])),
     ('aws s3 cp -', -1, set(['--no-guess-mime-type', '--dryrun',