+++ tests/unit/customizations/s3/test_utils.py
@@ -297,15 +297,19 @@ class TestScopedEventHandler(unittest.TestCase):
         session = mock.Mock()
         scoped = ScopedEventHandler(session, 'eventname', 'handler')
         with scoped:
-            session.register.assert_called_with('eventname', 'handler', None)
-        session.unregister.assert_called_with('eventname', 'handler', None)
+            session.register.assert_called_with('eventname', 'handler', None,
+                                                False)
+        session.unregister.assert_called_with('eventname', 'handler', None,
+                                              False)
 
     def test_scoped_session_unique(self):
         session = mock.Mock()
         scoped = ScopedEventHandler(session, 'eventname', 'handler', 'unique')
         with scoped:
-            session.register.assert_called_with('eventname', 'handler', 'unique')
-        session.unregister.assert_called_with('eventname', 'handler', 'unique')
+            session.register.assert_called_with('eventname', 'handler',
+                                                'unique', False)
+        session.unregister.assert_called_with('eventname', 'handler', 'unique',
+                                              False)
 
 
 class TestGetFileStat(unittest.TestCase):